# ENH, PERF: allocate memory as part of the array object for scalars