--- a/numpy/_core/src/multiarray/arrayobject.c
+++ b/numpy/_core/src/multiarray/arrayobject.c
@@ -426,26 +426,32 @@ array_dealloc(PyArrayObject *self)
                 PyErr_WriteUnraisable(NULL);
             }
         }
-        /* Deallocate data, if not a small allocation in the object */
-        if (fa->data != (void *)((char *)fa + Py_TYPE(self)->tp_basicsize)) {
-            if (fa->mem_handler == NULL) {
-                if (npy_thread_unsafe_state.warn_if_no_mem_policy) {
-                    char const *msg = "Trying to dealloc data, but a memory policy "
-                        "is not set. If you take ownership of the data, you must "
-                        "set a base owning the data (e.g. a PyCapsule).";
-                    WARN_IN_DEALLOC(PyExc_RuntimeWarning, msg);
-                }
-                // Guess at malloc/free ???
-                free(fa->data);
+        if (fa->mem_handler != NULL) {
+            size_t nbytes = PyArray_NBYTES(self);
+            if (nbytes == 0) {
+                nbytes = 1;
             }
-            else {
-                size_t nbytes = PyArray_NBYTES(self);
-                if (nbytes == 0) {
-                    nbytes = 1;
-                }
-                PyDataMem_UserFREE(fa->data, nbytes, fa->mem_handler);
-                Py_DECREF(fa->mem_handler);
+            PyDataMem_UserFREE(fa->data, nbytes, fa->mem_handler);
+            Py_DECREF(fa->mem_handler);
+        }
+        else if (PyArray_DATA(self) !=
+                 (void *)((char *)self + Py_TYPE(self)->tp_basicsize)) {
+            /*
+             * Without a handler, data should be on the instance; if not,
+             * someone did something arbitry, so warn and try free.
+             *
+             * TODO: strengthen this so we can refuse the temptation to guess?
+             * mem_handler = NULL should just signal no deallocation needed.
+             * If we do this, also remove the error path in array_setstate.
+             */
+            if (npy_thread_unsafe_state.warn_if_no_mem_policy) {
+                char const *msg = "Trying to dealloc data, but a memory policy "
+                    "is not set. If you take ownership of the data, you must "
+                    "set a base owning the data (e.g. a PyCapsule).";
+                WARN_IN_DEALLOC(PyExc_RuntimeWarning, msg);
             }
+            // Guess at malloc/free ???
+            free(fa->data);
         }
     }
 
@@ -1227,7 +1233,6 @@ NPY_NO_EXPORT PyTypeObject PyArray_Type = {
     PyVarObject_HEAD_INIT(NULL, 0)
     .tp_name = "numpy.ndarray",
     .tp_basicsize = sizeof(PyArrayObject_fields),
-    .tp_itemsize = 1,
     /* methods */
     .tp_dealloc = (destructor)array_dealloc,
     .tp_repr = (reprfunc)array_repr,