… numpy/_core/src/multiarray/shape.c