The overhead of allocating memory for array data becomes large when the inputs are scalars. So for those this PR allocates memory as part of the object using `tp_alloc`. (This solves the same problem #29876 aimed to solve, but arguably in a much nicer manner.)

Timings
```
%timeit np.empty(())
112->67 ns

a = 1.6
%timeit np.array(a)
134->89 ns

%timeit np.add(a, a)
527->413 ns (390 ns for np.float64)
```

In principle, this could be trivially extended to doing this generally when the required amount of memory is small.

One note: the above requires setting `tp_itemsize=1` for the array object, which means that always at least 1 extra byte gets allocated [1].  I verified that with the current struct, this does not increase the size of the object (it remains 96 bytes).  Alignment should be OK, since the last item on the current struct is a pointer.

[1] https://github.com/python/cpython/blob/37d16f7f620d79af7db1c00c8d638122af889bf9/Objects/typeobject.c#L2490-L2497