@mhvk 2025-10-05
Oops, I need to do better checking, `spin ipython` worked fine but `spin test` immediately breaks. Sigh.