@mhvk 2025-10-05
p.s. Am confused why the tests all fail when on my machine `spin test` works correctly.