@seberg 2025-10-05
> I was thinking that the allocator would be mostly used for very large arrays. But might it be needed in case someone, e.g., expects all data to be allocated in GPU memory or so? But wouldn't they then want the whole object there anyway?

Yes, precisely, someone might be asking for an alignment of 512bytes or so and any new array allocated within a context _and returned to the user_ should be expected to have this.
Arguably, we don't have to care for e.g. ufuncs (or in fact any array created from a scalar where a `copy=True` isn't done, I guess), but we don't have that information easily available here.

> p.s. Am confused why the tests all fail when on my machine spin test works correctly.

To be fair, a lot of tests are passing here also :).  Maybe try running with `PYTHONMALLOC=malloc_debug` or so, might make the failure obvious.