@mhvk 2025-10-05
Hmm, this is trickier than I hoped. Under python 3.11, I can get rid of the crashes by starting `PyArrayObject_fields` with `PyObject_VAR_HEAD`, but then multiple inheritance breaks:
```
class A(np.ndarray):
    pass

class B(np.ndarray):
    pass

class C(A, B):
    pass

TypeError: multiple bases have instance lay-out conflict
```
Under python 3.13, this all works fine.

I'll push the version with the change anyway, since it would seem to be correct, but it is unclear if I really can proceed like this...