@mhvk 2025-10-06
OK, I ended up just doing the allocation directly, instead of via `tp_alloc`. That works for both python versions, and keeps the basic size of the instances the same. I guess the non-so-great part is that this is hidden; suggestions for improvements welcome.

p.s. Still need to do the `mem_handler` check, but first want to be sure people are happy with proceeding along these lines...