@mhvk 2025-10-07
> I am fine with this as such, but I would like to have slightly more confidence about bare PyObject_malloc use here.

As mentioned in-line, that use comes straight from `PyType_GenericAlloc`, so is no change in that sense.

> There might also be an interesting thing whether this whole thing is just much faster on Python 3.14 or so due to the use of `mimalloc` (or is that only free-threading)?

I'll see if I can get 3.14rc3 to install...

> And in the end, I have to also think about `arr.resize` calls which might be weird but we can likely figure something out there (although need a release note, I wouldn't be surprised if someone is insane enough to replace NumPy array allocations under our feet...).

Oops, I thought I had looked carefully for other uses of the memory handler, but missed it because I was looking for `mem_handler` rather than `PyArray_HANDLER`. I fixed the resize (though noticed what I think is a bug in it; see #29887). Turns out there is another one in `textreading/rows.c`, though there the array is always allocated with `ndim>=1`, so it is unaffected. It does make me wonder if it wouldn't be better to have a fake'ish memory handler after all, though.