@mhvk 2025-10-07
OK, I now also added the `mem_handler == PyDataMem_DefaultHandler` check. As you said, it adds about 5 to 10 ns, which is a pity, but it still is a nice improvement on before.

Separately, you mentioned somewhere (I think) that the actual allocation functions might be sped up a bit. Were you thinking about the fact that every time the handler capsule is opened? It would indeed seem worthwhile to check whether it is the default handler and then avoid that pointer chasing...