@seberg 2025-10-07
> Were you thinking about the fact that every time the handler capsule is opened?

Right.  I had a start here with a slightly heavy handed approach to add a custom `PyCapsule_GetPointerUnchecked()` that assumes the pointer is the first thing stored in the capsule object (feels like plausible Python C-API, although maybe not on GraalPython).
(UFuncs also use a capsule right now for the errstate, so one more little thing.  But it was <5%, although I did most timings with `spin ipython` which does a debug build, which is terrible for these timings.)

It's not much, I was surprised to see that a try to avoid the _any_ allocation for scalar ufunc inputs made only a <10%, but then profiling isn't quite as simple as that.  (I should time this, I wonder if I barely notice it on the M1 in a simple `%timeit`)