@mhvk 2025-10-07
> Right. I had a start here with a slightly heavy handed approach to add a custom `PyCapsule_GetPointerUnchecked()` that assumes the pointer is the first thing stored in the capsule object (feels like plausible Python C-API, although maybe not on GraalPython).

I think one could be faster by just defining a little in-line function that has a shortcut, like
```
static inline PyDataMem_Hander*
_get_hander(PyObject *mem_handler_capsule) {
    if (mem_handler_capsule == PyDataMem_DefaultHandler) {
        return &default_handler;
    }
    else {
        return (PyDataMem_Handler *) PyCapsule_GetPointer(
            mem_handler_capsule, MEM_HANDLER_CAPSULE_NAME);
    }
}
```
and then in all the `PyDataMem` functions one would have
```
PyDataMem_Handler *handler = _get_handler(mem_handler);
```
which would be nicer regardless.

p.s. Am slightly sad that the handler isn't just supposed to take care of the tracking as well, i.e., that the handler itself contains pointers to `UserNEW`, etc. One could imagine a scenario where one just wants to dispense with that to reduce overhead, but that is not possible. Anyway, that ship has sailed, presumably.