@mhvk 2025-10-08
Hmm, MSVC does not like the zero-sized array, unfortunately. I went for the size-1 version, where that is not included in `tp_basicsize` (I do like the marker idea, explicit is better!).