@mhvk 2025-10-08
> Sorry, about Python versions. Without thinking it through: Yes, I can see doing such tricks only on Python versions that properly support it.

I checked and with python 3.12, defining an `itemsize` and using `tp_alloc(subtype, n_items)` works fine. I'm not sure that is necessarily better - it does add 8 bytes to every object for `ob_size`... And I'm not sure it would break all kinds of things if the array layout changed that way.

So, perhaps what's here is better. Do you think I should go for a separate `allocfunc`? It feels a bit cleaner, but also less explicit.