@seberg 2025-10-08
> it does add 8 bytes to every object for ob_size... And I'm not sure it would break all kinds of things if the array layout changed that way.

Yes, this is effectively impossible unfortunately.  (You would have to force every to recompile _and_ fix a lot of downstream projects that don't just use our headers directly.  So yeah, that is completely impossible, doing it would need a year long preparation at least.)